+++ R/grids.R
@@ -123,12 +123,10 @@ grid_regular.workflow <- function(x,
                                   levels = 3,
                                   original = TRUE,
                                   filter = NULL) {
-  grid_regular.parameters(
-    parameters(x),
-    ...,
-    levels = levels,
-    original = original,
-    filter = {{ filter }}
+  lifecycle::deprecate_stop(
+    when = "1.2.0",
+    what = "grid_regular.workflow()",
+    details = "Alternatively, first extract the parameter set with `extract_parameter_set_dials()`, then create the grid from that object."
   )
 }
 
@@ -247,12 +245,10 @@ grid_random.param <- function(x, ..., size = 5, original = TRUE, filter = NULL) 
 #' @export
 #' @rdname grid_regular
 grid_random.workflow <- function(x, ..., size = 5, original = TRUE, filter = NULL) {
-  grid_random.parameters(
-    parameters(x),
-    ...,
-    size = size,
-    original = original,
-    filter = {{ filter }}
+  lifecycle::deprecate_stop(
+    when = "1.2.0",
+    what = "grid_random.workflow()",
+    details = "Alternatively, first extract the parameter set with `extract_parameter_set_dials()`, then create the grid from that object."
   )
 }