+++ R/space_filling.R
@@ -111,13 +111,10 @@ grid_max_entropy.param <- function(x, ..., size = 3, original = TRUE,
 #' @rdname grid_max_entropy
 grid_max_entropy.workflow <- function(x, ..., size = 3, original = TRUE,
                                       variogram_range = 0.5, iter = 1000) {
-  grid_max_entropy.parameters(
-    parameters(x),
-    ...,
-    size = size,
-    original = original,
-    variogram_range = variogram_range,
-    iter = iter
+  lifecycle::deprecate_stop(
+    when = "1.2.0",
+    what = "grid_max_entropy.workflow()",
+    details = "Alternatively, first extract the parameter set with `extract_parameter_set_dials()`, then create the grid from that object."
   )
 }
 
@@ -210,7 +207,11 @@ grid_latin_hypercube.param <- function(x, ..., size = 3, original = TRUE) {
 #' @export
 #' @rdname grid_max_entropy
 grid_latin_hypercube.workflow <- function(x, ..., size = 3, original = TRUE) {
-  grid_latin_hypercube.parameters(parameters(x), ..., size = size, original = original)
+  lifecycle::deprecate_stop(
+    when = "1.2.0",
+    what = "grid_latin_hypercube.workflow()",
+    details = "Alternatively, first extract the parameter set with `extract_parameter_set_dials()`, then create the grid from that object."
+  )
 }